Examples/API Management/Data upload client/requirements.txt
azure-storage-blob==12.12.0
Flask==2.0.2
requests==2.25.1
urllib3==1.26.4
validate-email==1.3
xmltodict==0.13.0